Use 64bit block numbers to handle LBA48
Also, remove extraneous len parameter from ll_rw_block

+++ include/scaraOS/types.h
@@ -2,9 +2,10 @@
 #define _SCARAOS_TYPES_H
 
 #include <arch/types.h>
+
 /* Kernel types */
 typedef uint32_t pid_t;
-typedef uint32_t block_t;
+typedef uint64_t block_t;
 typedef uint32_t ino_t;
 typedef uint32_t uid_t;
 typedef uint32_t gid_t;
